ticket:1573 - Fixed all errors, that were preventing compilation, and added few new compilation flags
Summary: ticket:1573 - Fixed all errors, that were preventing compilation, and added few new compilation flags

Test Plan: Build, compile, run and perform queries.

Reviewers: dmisev

Reviewed By: dmisev

Differential Revision: http://codereview.rasdaman.org/D894

+++ applications/rasql/rasql.cc
@@ -1164,7 +1164,7 @@ void doStuff(__attribute__((unused)) int argc, __attribute__((unused)) char **ar
             }
             fseek(fileD, 0, SEEK_SET);
             size_t rsize = fread(fileContents.get(), 1, static_cast<size_t>(size), fileD);
-            if (rsize != size)
+            if (static_cast<long>(rsize) != size)
             {
                 BLERROR << "failed, read only " << rsize << " bytes of " << size << " bytes.\n";
                 throw RasqlError(FILEREADERROR);
